+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * add support for issue comments [#71](https://github.com/softprops/hubcaps/pull/71)
 * add support for repo teams, and iterators over all pages of both org and repo teams [#72](https://github.com/softprops/hubcaps/pull/72)
 * add team permission
+* add iter support repo and pull request listing
 
 # 0.3.1
 * fix order of Iter traversal